Fixes in changed messages for int to long

+++ src/main/java/io/swagger/client/model/To.java
@@ -73,7 +73,7 @@ public class To {
     this.resource = resource;
   }
 
-  public To resourceId(Integer resourceId) {
+  public To resourceId(Long resourceId) {
     this.resourceId = resourceId;
     return this;
   }
@@ -83,11 +83,11 @@ public class To {
    * @return resourceId
   **/
   @ApiModelProperty(value = "")
-  public Integer getResourceId() {
+  public Long getResourceId() {
     return resourceId;
   }
 
-  public void setResourceId(Integer resourceId) {
+  public void setResourceId(Long resourceId) {
     this.resourceId = resourceId;
   }
